add scan-errors logs
